Reorganise file hierarchy/structure
